% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evt_cdf.R
\name{evt_cdf}
\alias{evt_cdf}
\title{evt_cdf}
\usage{
evt_cdf(
  n = 100,
  t = seq(0, 1, by = 0.01),
  remp = function(x) stats::runif(x),
  ptrue = function(x) stats::punif(x),
  ...
)
}
\arguments{
\item{n}{Number of samples}

\item{t}{Layout x-axis grid with a function such as t = seq(0, 1, by = .01)}

\item{remp}{Take random samples from a distribution such as remp = runif(n)}

\item{ptrue}{Get theoretical cumulative distribution with a function like ptrue = punif(t)}

\item{...}{Add extra arguments such as plot title and axis labels}
}
\value{
A plot with the empirical and true  cdfs
}
\description{
Empirical versus theoretical cumulative distribution plot.
}
\details{
Based on the OCW example: https://www.youtube.com/watch?v=8fEETchU7NM&t=1458s
}
\examples{
evt_cdf()
evt_cdf(remp = function(x)rnorm(x),
        ptrue = function(x)pnorm(x))
}
