% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simq.R
\name{simq}
\alias{simq}
\title{simq}
\usage{
simq(
  estimator = function(x) mean(x),
  reps = 30,
  sims = 10000,
  frame = 1:100,
  ref_samp,
  lowp = 0.025,
  highp = 0.975,
  replace = TRUE,
  main = "Distribution of Samples",
  xlab = "Sample Estimates",
  ylab = "Frequency",
  breaks = 20
)
}
\arguments{
\item{estimator}{Anonymous function defining estimator of interest}

\item{reps}{Number of samples per iteration of simulation}

\item{sims}{Number of iteration in simulation}

\item{frame}{Sampling frame from which to draw sample}

\item{ref_samp}{A reference sample (reference sample estimate is blue abline
in histogram)}

\item{lowp}{Lower percentile in stats::quantile(.,probs=c(lowp,highp))
function (left red abline in histogram)}

\item{highp}{Higher percentile in stats::quantile(.,probs=c(lowp,highp))
function (right red abline in histogram)}

\item{replace}{Parameter for base::sample}

\item{main}{Histogram title}

\item{xlab}{Histogram x-axis label}

\item{ylab}{Histogram y-axis label}

\item{breaks}{Histogram breakpoints}
}
\value{
A list with two objecs: 1) Return the requested quantiles from "sims"
iterations of the simulation and 2) Histogram with reference ablines with references for lowp, highp, and ref_samp over "sims" iterations of the simulation
}
\description{
Simulate quantiles for a sample estimator
}
\examples{
#Test scores from the new test administered over the past 3 years is not
#available but the mean (75) and sd (14) is known
sim_scores <- as.vector(rnorm(1000, 75, 14))

#Practice test scores for low performing math class
practice_test <- as.vector(rnorm(30, 68, 7))

#See how low performers compare to past performers

#Run simulation
getqs <- simq(frame = sim_scores, ref_samp = practice_test ,
              main = "Distribution of Practice Tests", xlab = "Sample Means")

#Get requested quantiles
getqs$quantiles

#Compare practice test to distribution of past tests
getqs$histogram
}
