% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simtest.R
\name{simtest}
\alias{simtest}
\title{SimTest}
\usage{
simtest(
  test = function(x) sum(x) >= 3,
  on = c(0, 1),
  reps = 10,
  sims = 10^2,
  replace = TRUE
)
}
\arguments{
\item{test}{An anonymous test function}

\item{on}{A vector of values to sample from}

\item{reps}{The number of samples for each iteration of test}

\item{sims}{The number of test iterations}

\item{replace}{Replace argument from the base R function: sample(x, size, replace = FALSE, prob = NULL)}
}
\value{
A vector with test results for "sims" iterations of the test
}
\description{
Use Monte Carlo Simulation on anonymous test function
}
\examples{
#The default settings of simtest checks whether reps = 10 samples from
#"on = c(0, 1)" returns a value of "1" three or more times.  This can be thought of
#as a test for whether flipping a coin 10 times returns heads three or more times.
#Example and function inspired from A Beginner's Guide to Monte Carlo Simulations
#presented by Dan Uehara for UseR Oslo: https://www.youtube.com/watch?v=g2Uu9m0IlFE
coinflips <- simtest()
mean(coinflips)
}
